pyproject.toml
[tool.poetry]
name = "eigenface"
version = "0.1.0"
description = ""
authors = ["Summerkirakira <summerkirakira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pillow = "^10.0.1"
numpy = "^1.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

